Move check_for_snap from activator to raws.
check_for_snap is much more tightly coupled to the raws code (paths,
parsing, etc.) than to the activator (bucket access). The activator is
still responsible for initializing the S3 client.
